requirements.txt
requests
python-dotenv
azure-ai-textanalytics
